+++ examples/sam_tutorial.jl
@@ -108,8 +108,7 @@ vsm_set = VortexStepMethod.VSMSettings(
     data_prefix=false)
 vsm_wing = VortexStepMethod.Wing(vsm_set)
 vsm_aero = BodyAerodynamics([vsm_wing])
-vsm_solver = Solver(vsm_aero;
-    solver_type=NONLIN, atol=2e-8, rtol=2e-8)
+vsm_solver = Solver(vsm_aero, vsm_set)
 wings = [SymbolicAWEModels.Wing(1, vsm_aero, vsm_wing,
     vsm_solver, [], I(3), [0.5, 0, set.l_tether + 6])]